backend/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
api==0.0.7
asgiref==3.8.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==4.1.13
django-cors-headers==4.6.0
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.7.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
nose==1.3.7
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pefile==2023.2.7
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pymongo==3.12.3
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post03
requests==2.32.3
six==1.16.0
sqlparse==0.2.4
tqdm==4.67.0
zipp==3.20.2
